content of publisher-monitor moved into publisher-commons
