+++ pom.xml
@@ -93,6 +93,11 @@
 	<dependencyManagement>
 		<dependencies>
 			<dependency>
+	    		<groupId>org.aspectj</groupId>
+	    		<artifactId>aspectjrt</artifactId>
+	    		<version>1.8.0</version>
+	    	</dependency>
+			<dependency>
 				<groupId>com.h2database</groupId>
 				<artifactId>h2</artifactId>
 				<version>${h2.version}</version>			
@@ -408,8 +413,7 @@
 
 	<modules>
 		<module>publisher-commons</module>
-		<module>publisher-provider</module>		
-		<module>publisher-monitor</module>		
+		<module>publisher-provider</module>
 		<module>publisher-database</module>
 		<module>publisher-service</module>
 		<module>publisher-domain</module>
