+++ publisher-commons/pom.xml
@@ -30,11 +30,7 @@
     	<dependency>
             <groupId>com.typesafe.akka</groupId>
             <artifactId>akka-actor_${scala.abi.version}</artifactId>            
-        </dependency>
-    	<dependency>
-    		<groupId>nl.idgis.publisher</groupId>
-    		<artifactId>publisher-monitor</artifactId>    		
-    	</dependency>
+        </dependency>    	
 		<dependency>
 			<groupId>org.apache.commons</groupId>
 			<artifactId>commons-lang3</artifactId>			
@@ -55,6 +51,10 @@
 			<groupId>nl.idgis.publisher</groupId>
 			<artifactId>publisher-test</artifactId>
 		</dependency>
+		<dependency>
+    		<groupId>org.aspectj</groupId>
+    		<artifactId>aspectjrt</artifactId>    		
+    	</dependency>
 		<dependency>
 			<groupId>com.h2database</groupId>
 			<artifactId>h2</artifactId>
