+++ publisher-commons/src/main/java/nl/idgis/publisher/monitor/MonitorAspect.java
