+++ publisher-commons/src/main/java/nl/idgis/publisher/monitor/messages/ResourceDestroyed.java
