+++ publisher-commons/src/main/java/nl/idgis/publisher/tree/Node.java
@@ -1,4 +1,4 @@
-package nl.idgis.publisher.monitor.messages;
+package nl.idgis.publisher.tree;
 
 public abstract class Node {
