+++ publisher-commons/src/main/resources/META-INF/aop.xml
