--- publisher-monitor/pom.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    
-    <artifactId>publisher-monitor</artifactId>
-    
-    <parent>
-        <groupId>nl.idgis.publisher</groupId>
-        <artifactId>publisher</artifactId>
-        <version>0.0.1-SNAPSHOT</version>
-    </parent>
-    
-    <dependencies>
-    	<dependency>
-            <groupId>com.typesafe.akka</groupId>
-            <artifactId>akka-actor_${scala.abi.version}</artifactId>            
-        </dependency>
-    	<dependency>
-    		<groupId>org.aspectj</groupId>
-    		<artifactId>aspectjrt</artifactId>
-    		<version>1.8.0</version>
-    	</dependency>
-    </dependencies>
-    
-</project>
